data_analysis_agent/requirement.txt
python>=3.10
langchain>=0.2.0
langchain-community>=0.2.0
langchain-core>=0.2.0
langchain-deepseek>=0.1.0
langchain-openai>=0.1.0
langgraph>=0.1.0
fastapi>=0.110.0
uvicorn>=0.29.0
pydantic>=2.0.0
mysql-connector-python>=8.2.0
requests>=2.31.0
pypdf>=4.0.0
openai>=1.0.0